Types no set in GetAqcuiredData

+++ Andor/andorSDK.pyx
@@ -136,10 +136,10 @@ class Andor:
 
     def GetAcquiredData(self, width, height):
         cdef unsigned int size = width*height
-        cdef array.array data = array.array('i', np.zeros(size,dtype=np.int))
+        cdef array.array data = array.array('i', np.zeros(size,dtype=np.int32))
         error = lib.GetAcquiredData(data.data.as_ints, size)
         self.verbose(error, "GetAcquiredData")
-        return np.array(data).reshape((height,width)).transpose()
+        return np.array(data,dtype=np.float64).reshape((height,width)).transpose()
         #return data
 
     def GetPixelSize(self):
